Improved: Actions/Filters classes are separated into more descriptive classes.
