+++ src/AdminBar.php
@@ -1,0 +1,109 @@
+<?php
+/**
+ * Plausible Analytics | Admin
+ */
+
+namespace Plausible\Analytics\WP;
+
+class AdminBar {
+	/**
+	 * Build class.
+	 */
+	public function __construct() {
+		$this->init();
+	}
+
+	/**
+	 * Action/filter hooks.
+	 *
+	 * @return void
+	 */
+	private function init() {
+		add_action( 'admin_bar_menu', [ $this, 'admin_bar_node' ], 100 );
+	}
+
+
+	/**
+	 * Create admin bar nodes.
+	 *
+	 * @param \WP_Admin_Bar $admin_bar Admin bar object.
+	 *
+	 * @return void
+	 * @since  1.3.0
+	 * @access public
+	 *
+	 */
+	public function admin_bar_node( $admin_bar ) {
+		$disable = ! empty( Helpers::get_settings()['disable_toolbar_menu'] );
+
+		if ( $disable ) {
+			return; // @codeCoverageIgnore
+		}
+
+		$settings     = Helpers::get_settings();
+		$current_user = wp_get_current_user();
+
+		$has_access             = false;
+		$user_roles_have_access = array_merge(
+			[ 'administrator' ],
+			$settings['expand_dashboard_access'] ?? []
+		);
+
+		foreach ( $current_user->roles as $role ) {
+			if ( in_array( $role, $user_roles_have_access, true ) ) {
+				$has_access = true;
+				break;
+			}
+		}
+
+		if ( ! $has_access ) {
+			return;
+		}
+
+		// Add main admin bar node.
+		$args[] = [
+			'id'    => 'plausible-analytics',
+			'title' => 'Plausible Analytics',
+		];
+
+		if ( ! empty( $settings['enable_analytics_dashboard'] ) || ( ! empty( $settings['self_hosted_domain'] ) && ! empty( $settings['self_hosted_shared_link'] ) ) ) {
+			$args[] = [
+				'id'     => 'view-analytics',
+				'title'  => esc_html__( 'View Analytics', 'plausible-analytics' ),
+				'href'   => admin_url( 'index.php?page=plausible_analytics_statistics' ),
+				'parent' => 'plausible-analytics',
+			];
+
+			// Add link to individual page stats.
+			if ( is_singular() ) {
+				global $post;
+				$uri = wp_make_link_relative( get_permalink( $post->ID ) );
+
+				$args[] = [
+					'id'     => 'view-page-analytics',
+					'title'  => esc_html__( 'View Page Analytics', 'plausible-analytics' ),
+					'href'   => add_query_arg(
+						'page-url',
+						is_home() ? '' : $uri,
+						admin_url( 'index.php?page=plausible_analytics_statistics' )
+					),
+					'parent' => 'plausible-analytics',
+				];
+			}
+		}
+
+		// Add link to Plausible Settings page.
+		if ( current_user_can( 'manage_options' ) ) {
+			$args[] = [
+				'id'     => 'settings',
+				'title'  => esc_html__( 'Settings', 'plausible-analytics' ),
+				'href'   => admin_url( 'options-general.php?page=plausible_analytics' ),
+				'parent' => 'plausible-analytics',
+			];
+		}
+
+		foreach ( $args as $arg ) {
+			$admin_bar->add_node( $arg );
+		}
+	}
+}
